Continued adding some core functionality for image processing and ROI management
